requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dynaconf==3.2.11
fastapi==0.116.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
iso8601==2.1.0
kombu==5.5.4
loguru==0.7.3
packaging==25.0
priority==2.0.0
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
win32_setctime==1.2.0
wsproto==1.2.0
